+++ src/app/constants/api.ts
@@ -11,3 +11,11 @@ export const OPTION = '/option';
 export const QINIU = '/qiniu';
 
 export const ANNOUNCEMENT = '/announcement';
+
+export const CATEGORY = '/category';
+
+export const ARTICLE = '/article';
+
+export const COMMENT = '/comment';
+
+export const TAG = '/tag';