+++ src/app/pages/demo/tables/components/basicTables/basicTables.component.ts
@@ -8,6 +8,5 @@ import { Component, ViewEncapsulation } from '@angular/core';
 })
 export class BasicTablesComponent {
 
-  constructor() {
-  }
+  constructor() {}
 }