+++ src/app/pages/demo/ui/ui.routing.ts
@@ -1,6 +1,6 @@
-import { Routes, RouterModule }  from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 
-import { Ui } from './ui.component';
+import { UiComponent } from './ui.component';
 import { Buttons } from './components/buttons/buttons.component';
 import { Grid } from './components/grid/grid.component';
 import { Icons } from './components/icons/icons.component';
@@ -11,7 +11,7 @@ import { Typography } from './components/typography/typography.component';
 const routes: Routes = [
   {
     path: '',
-    component: Ui,
+    component: UiComponent,
     children: [
       { path: 'buttons', component: Buttons },
       { path: 'grid', component: Grid },